Move platform.h to /src/.
