+++ src/control_loop.cpp
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the Apache-2.0 license, see LICENSE
 #include "control_loop.h"
 
-#include <franka/platform_type.h>
+#include "platform_type.h"
 #ifdef LIBFRANKA_WINDOWS
 #include <Windows.h>
 #else
