+++ src/platform_type.h
